+++ include/terra/aescrypt/engine/encryptor.h
@@ -122,7 +122,7 @@ class Encryptor
         std::string instance;
         bool active;
         bool cancelled;
-        std::uint8_t modulo;
+        std::uint8_t read_length;
         std::size_t octets_consumed;
         std::size_t progress_octets;
         std::condition_variable cv;